% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwHMM-package.R
\docType{data}
\name{fit}
\alias{fit}
\title{crwHMM example fit object}
\format{.RData}
\description{
Example crwHMM fit object, using ellie example data and the
following call: fit <- fit_ssm(ellie, model="rw", time.step=24). This example
fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.
}
\keyword{data}
