% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.crwHMM.R
\name{print.crwHMM}
\alias{print.crwHMM}
\title{print \code{crwHMM} object summary information}
\usage{
\method{print}{crwHMM}(x, ...)
}
\arguments{
\item{x}{a \code{ctrw} fit object}

\item{...}{unused. For compatibility with the generic method.}
}
\description{
print \code{crwHMM} object summary information
}
\examples{
## see summary fit output
## load example crwHMM fit object (to save time)
data(fit)
fit

}
